[ADD] stock_request_kanban
