+++ stock_request/models/__init__.py
@@ -1,3 +1,4 @@
+from . import stock_request_abstract
 from . import stock_request
 from . import stock_request_allocation
 from . import stock_request_order
