+++ stock_request/views/res_config_settings_views.xml
@@ -46,6 +46,21 @@
                             </div>
                         </div>
                     </div>
+                    <h2>Kanban</h2>
+                    <div class="row mt16 o_settings_container" id="stock_request_purchase">
+                        <div class="col-xs-12 col-md-6 o_setting_box">
+                            <div class="o_setting_left_pane">
+                                <field name="module_stock_request_kanban"/>
+                            </div>
+                            <div class="o_setting_right_pane">
+                                <label string="Enable Stock Requests Kanban cards"
+                                       for="module_stock_request_purchase"/>
+                                <div class="text-muted">
+                                    Use Kanban cards for consumable products
+                                </div>
+                            </div>
+                        </div>
+                    </div>
                 </div>
             </xpath>
         </field>
